Changed the position of toggle icon
