+++ src/components/navBar/index.tsx
@@ -1,9 +1,15 @@
 import Image from 'next/image';
+import { FC } from 'react';
+import Toggle from '../dark-theme/toggler';
 import styles from '@/components/navBar/navBar.module.scss';
+
+interface Props {
+  theme?: any,
+}
 
 const RDSLogo = '/RDSLogo.png';
 
-const NavBar = () => (
+const NavBar: FC<Props> = ({ theme }) => (
   <nav className={styles.navBar}>
     <ul>
       <li>
@@ -14,6 +20,7 @@ const NavBar = () => (
       <li><a href="https://members.realdevsquad.com/">Members</a></li>
       <li><a href="https://crypto.realdevsquad.com/">Crypto</a></li>
       <li><a className={styles.active} href="https://status.realdevsquad.com/">Status</a></li>
+      <Toggle themeMode={theme} style={styles.iconImage} />
     </ul>
   </nav>
 );
